since pyqt5 version something, qprogressbar accepts only int type, @dkriegner please acknowledge

+++ matr1x/control/util.py
@@ -110,7 +110,7 @@ def connectDictValueToDisplay(connDict):
             connDict[key][0].valueChanged[str].connect(
                 connDict[key][1][1].setText)
         elif type(connDict[key][1][1]) == QProgressBar:
-            connDict[key][0].valueChanged[float].connect(
+            connDict[key][0].valueChanged[int].connect(
                 connDict[key][1][1].setValue)
         elif type(connDict[key][1][1]) == QComboBox:
             connDict[key][0].valueChanged[int].connect(
